quick fixes to filtering
filtering was converted to a binary image and then filtered as a binary
image. This caused a lot of data to be thrown away when filtering.

Because change causes issues with detection in the computer generated
fonts I made another function identical to the original and called that
from the trainer instead.